Move single hotel map to map/singleHotel
