+++ app/hotels/src/map/singleHotel/MapView.js
