+++ app/hotels/src/navigation/SingleHotelMapNavigationScreen.js
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { withMappedNavigationAndConfigProps } from 'react-navigation-props-mapper';
 
 import type { AvailableHotelSearchInput } from '../singleHotel/AvailableHotelSearchInput';
-import SingleHotelMap from '../singleHotel/map/SingleHotelMapScreen';
+import SingleHotelMap from '../map/singleHotel/SingleHotelMapScreen';
 
 type Props = {
   ...AvailableHotelSearchInput,
